+++ app/components/metadata-panel.tsx
@@ -175,7 +175,7 @@ const useOtherStyles = makeStyles<Theme>((theme) => {
         borderBottom: "none",
       },
     },
-    openDimension: {
+    openComponent: {
       minHeight: 0,
       verticalAlign: "baseline",
       padding: 0,
@@ -211,18 +211,24 @@ export const OpenMetadataPanelWrapper = ({
   component,
 }: {
   children: ReactNode;
-  component: Component;
+  component?: Component;
 }) => {
   const classes = useOtherStyles();
-  const { openDimension } = useMetadataPanelStoreActions();
+  const { openComponent, setOpen, setActiveSection } =
+    useMetadataPanelStoreActions();
   const handleClick = useEvent((e: React.MouseEvent) => {
     e.stopPropagation();
-    openDimension(component);
+    if (component) {
+      openComponent(component);
+    } else {
+      setActiveSection("general");
+      setOpen(true);
+    }
   });
 
   return (
     <Button
-      className={classes.openDimension}
+      className={classes.openComponent}
       variant="text"
       size="small"
       onClick={handleClick}
@@ -473,9 +479,9 @@ const DataPanel = ({
   const locale = useLocale();
   const classes = useOtherStyles();
   const selectedDimension = useMetadataPanelStore(
-    (state) => state.selectedDimension
+    (state) => state.selectedComponent
   );
-  const { setSelectedDimension, clearSelectedDimension } =
+  const { setSelectedComponent, clearSelectedComponent } =
     useMetadataPanelStoreActions();
   const [inputValue, setInputValue] = useState("");
   const { options, groupedComponents } = useMemo(() => {
@@ -552,7 +558,7 @@ const DataPanel = ({
         {selectedDimension ? (
           <MotionBox key="dimension-selected" {...animationProps}>
             <BackButton
-              onClick={() => clearSelectedDimension()}
+              onClick={() => clearSelectedComponent()}
               sx={{ color: "primary.main" }}
             >
               <Trans id="button.back">Back</Trans>
@@ -574,7 +580,7 @@ const DataPanel = ({
             <Autocomplete
               className={classes.search}
               disablePortal
-              onChange={(_, v) => v && setSelectedDimension(v.value)}
+              onChange={(_, v) => v && setSelectedComponent(v.value)}
               inputValue={inputValue}
               onInputChange={(_, v) => setInputValue(v.toLowerCase())}
               options={sortedOptions}
@@ -686,7 +692,7 @@ const ComponentTabPanel = ({
   cubeIri?: string;
 }) => {
   const classes = useOtherStyles();
-  const { setSelectedDimension } = useMetadataPanelStoreActions();
+  const { setSelectedComponent } = useMetadataPanelStoreActions();
   const label = useMemo(
     () => getComponentLabel(component, { cubeIri }),
     [cubeIri, component]
@@ -745,9 +751,9 @@ const ComponentTabPanel = ({
 
   const handleClick = useCallback(() => {
     if (!expanded) {
-      setSelectedDimension(component);
+      setSelectedComponent(component);
     }
-  }, [expanded, component, setSelectedDimension]);
+  }, [expanded, component, setSelectedComponent]);
 
   return (
     <div>