+++ CHANGELOG.md
@@ -11,6 +11,13 @@ You can also check the
 
 # Unreleased
 
+- Features
+  - It's now possible to label individual chart tabs
+  - Time is now displayed in addition to date in `Last edit` column in user
+    profile
+  - Free canvas layout option now includes a 3-column layout, depending on the
+    screen width
+  - Added a new popover login menu
 - Fixes
   - Map is now correctly centered after copying a chart or switching to layout
     mode
@@ -20,7 +27,10 @@ You can also check the
   - Renaming of charts through user profile now works correctly
   - Manually entering dates in date pickers works correctly again
   - Improved scrolling behavior in the chart tabs
+- Style
   - Aligned editor and layouting page layouts
+  - Removed dimension selection from modal when merging cubes
+  - Updated styles of confirmation dialog
 - Maintenance
   - Updated cube-hierarchy-query library to 3.0.0